+++ NEWS.md
@@ -1,7 +1,31 @@
+# antaresEditObject 0.7.0 (development)
+
+> Second-member coupling constraint scenarios
+
+NEW FEATURES (Antares v8.7, cf. Antares v8.7 changelog) :  
+
+* `createBindingConstraint()` / `createBindingConstraintBulk()` 
+  - New parameters `group`  
+  - Parameter `values` is now list of `data.frame` 
+  
+* `editBindingConstraint()` 
+  - New parameters `group`  
+  - Parameter `values` is now list of `data.frame`
+  
+* `removeBindingConstraint()` can now delete coupling constraints from the `group` parameter.
+* `scenarioBuilder()` has 3 new parameters dedicated to use with binding constraints.
+* `updateGeneralSettings()` adds coupling constraints to the `scenariobuilder.dat` file.
+ 
+### Breaking changes  :  
+
+* `createBindingConstraint()` is available with **offset** parameter in API mode 
+
+
 # antaresEditObject 0.6.4 (development)
 
 BREAKING CHANGES :
 * Add UTF-8 encoding argument in `.getJobs()`
+* Unit tests no longer call the study in the antaresRead package for versions > 8.0.0
 
 BUGFIXES :
 * `createArea()`/`editArea()` : in API mode, split data in nodalOptimization argument to write it in the expected files
@@ -62,10 +86,10 @@ BUGFIXES :
 # antaresEditObject 0.6.1
 * `writeInputTS()` allows the user to set a link with the separator ' - ' (ex: 'area1 - area2')
 
-
 BUGFIXES : 
 
 * Error CRAN CHECKS (fix issue #115)
+
 
 # antaresEditObject 0.6.0
 
@@ -97,7 +121,6 @@ NEW FEATURES :
 * Add `writeIniHydro()` function to make easier the edition of the input/hydro/hydro.ini file
 * Call `writeIniHydro()` in `createArea()` and `removeArea()`
 * `removeArea()` removes only expected files in links directory
-
 
 
 ### Breaking changes